requirements.txt
Flask==2.0.1
requests==2.25.1
core==1.0.1
Pillow==8.3.0
